new baseclass implementation
